Iterations instead of recursion and resume after failure
The memeSHARK can now uses more iterations instead of recursion for the traversal of the git graph. This should prevent problems with the maximum recursion depth. Additionally, the memeSHARK was extended such that commits that were already processed are skipped to allow resuming after breaks or failures (in most cases).